Rename eqntools to reltools
