--- sympy/relation/eqntools/__init__.py
@@ -1,7 +1,0 @@
-"""
-Module to implement equation manipulation algorithm.
-"""
-
-__all__ = ['eqnsimp']
-
-from .simplify import eqnsimp
