+++ sympy/relation/reltools/__init__.py
@@ -1,0 +1,7 @@
+"""
+Module to implement symbolic manipulation algorithm of relations.
+"""
+
+__all__ = ['eqnsimp']
+
+from .simplify import eqnsimp
