+++ sympy/relation/reltools/simplify.py
